welcomeVC setup
